JS functionality test
In order to build future functionality of user input to actually play this keyboard, I wanted to quickly test out user input with JS. For this test, by clicking on a specific note I was able to trigger an audio file to play. Success!

In the future, I'd like to have the user able to switch between the following sounds:
1. my actual acoustic piano that I will sample notes at soft, medium, and loud velocities
2. basic synth waveforms
3. other fun samples