Video embed, hero adjustment
I made the hero section more appealing and obvious, and made a number of cosmetic changes throughout.

I designated sections as hero, intro, features, video, and how it works.

I also embedded a video of one of the most famous genies, though in practice it would be better to have a video of the product/service.

next up is to tackle the "how it works" section.