scroll behavior and content bg
There's now better handling when using the links in the nav bar:smooth scrolling, and the title you scroll to is no longer obscured by the header.

Slight bg gradient adjustment to draw the eye in to the text more.

Also, I'm deciding to present the text on a semi-transparent background to make it more cohesive.