image gather and UI improvements
I gathered an image to represent each of the 10 planets and continued to improve the way the information is displayed.
Using one image as a stand-in, I was able to see how this might look in the final version.

Next up is to get the corresponding images to switch on the button click.