dynamic bg and favicon swap
I created a custom context to import all the images into an object rather than use individual import statements at the top of the component.

With this object, I used inline styling to dynamically update the element's background image when the planet's data shows.

I also updated the favicon to the rebel logo to match the project theme.