Refactor fetching logic
by moving the fetch data logic into the parent component of the one that uses it, when that child changes state it doesn't require a re-fetching of all the data again.
Instead, you just pull the data one time and mess with it to heart's content. In this case, that means applying some animation fades to the button click event.