Dysfunctional version
I wanted to put this here as a reminder of how *not* to structure a project in react. I see now how this is not how react was designed to be used; I was trying to mutate the props that were passed to the child instead of simply representing its state. the commit after this should be a rectification.