Fully Operational Randomization
I got it!!!

Finally. I was able to randomly cycle through the 10 objects in the fetched data array without repeating a given object until all 10 had been visited. Once all objects had been visited, the same unique visitation behavior continues. This required the
creation of an array (via the useRef hook to persist data across renders) that held values which correspond to indices of the passed data props.

Further explanation:
When a given planet is shown, its corresponding index is the current state. That index is used to index the data props, showing that planet.

When the user clicks the "visit new planet" button, so long as there are remaining values in the array of indices, the new state of index will become one of those values randomly, which then is used as described above to select the planet from the data props. Also, that value is removed from the array of indices. When that array is emptied, on click there is a check to see if that's true, in which case the array is restored to carry indices for all the data props.

One other note: there's a recursive check to ensure that when the array is restored, the index that emptied the array can not be the index that is first gotten from the restored array.
ex: if 4 was the final value to empty out, 4 will not be called first.